requirements.txt
flask
flask-sqlalchemy
flask-security
flask-Admin
mysqlclient
flask-marshmallow
bcrypt
WTForms
Flask-WTF
email-validator
mysql-connector-python
